requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
crispy-bootstrap5==0.3.1
Django==3.2
django-crispy-forms==1.11.2
django-multiselectfield==0.1.12
django-resized==0.3.11
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
